BASH unstack and minor issues

+++ bash/clone.sh
@@ -7,7 +7,12 @@ create_branch_directory() {
 clone_project() {
     destination=$(get_devstack)
     su -l $USER_NAME -c "git clone -b $DEVSTACK_BRANCH $DEVSTACK_REPO_URL $destination"
-    su -l $USER_NAME -c "cp $destination/local.sh $destination/local.sh.bkp"
+    if [ -f "$destination/local.sh" ]; then
+        su -l $USER_NAME -c "cp $destination/local.sh $destination/local.sh.bkp"
+    else
+        su -l $USER_NAME -c "touch $destination/local.sh"
+        su -l $USER_NAME -c "touch $destination/local.sh.bkp"
+    fi
 }
